Move business logic from controller to service
